attempting to finish level 1
i'm realizing that i'm creating a lot of scripts that seem like they could probably be combined into something more elegant but oh well.
my commit descriptions suck now. oh well though

+++ Assets/Scripts/Obstacles/DoorController.cs
@@ -7,7 +7,7 @@ public class DoorController : MonoBehaviour
     public bool isOpen = false;
     private Rigidbody2D body;
     private Vector2 startPos;
-    [SerializeField] private int speed;
+    [SerializeField] private float speed;
 
     private void Start()
     {
